refactor(renderer): consume the binding as a normal module; drop the build tag
Both changes remove bespoke machinery now that upstream has stabilised.

## Standard module dependency

The Go binding was consumed through a `replace` directive pointing at
/vendor-maplibre-go, a directory the Docker build copied out of the FFI
checkout. Upstream publishes bindings/go as a real module, so it is now an
ordinary requirement resolved from the proxy and checksummed in go.sum:

    github.com/maplibre/maplibre-native-ffi/bindings/go v0.0.0-20260728081959-91ecc9204624

There are no semver tags upstream yet, so this is a pseudo-version. That
turns out to be useful: its suffix is the upstream commit, so the build can
verify the binding and the C ABI it calls came from the same revision —
the property the vendored copy gave implicitly. rampardos-build now asserts
go.mod's pseudo-version against MLN_FFI_REV and fails with the exact
`go get` command to fix it. MLN_FFI_REV is promoted to a global ARG so both
stages read one value. Assertion verified to pass on match and fail on
drift.

Drops the vendor copy and its COPY from the image.

## Drop the mln_ffi build tag

The renderer needs EGL and links libmaplibre-native-c.so, so it is
Linux-only. That was expressed with a custom `mln_ffi` tag, which had two
problems: it could be forgotten, silently producing a binary with no
renderer, and it hid gopool.go from `go build ./...` — which is exactly how
deleting nodepool.go removed three helpers gopool.go used without any local
build noticing.

Now expressed the standard Go way, by filename:

    gopool.go       -> gopool_linux.go
    gopool_stub.go  -> gopool_other.go   (//go:build !linux)
    egl_linux.go    -> tag line removed, suffix already constrained it
    tests           -> *_linux_test.go

A Linux build always contains the real renderer; there is nothing to
forget. The stub keeps macOS compiling and testing everything else, which
matters because development happens on macOS while deployment is a Linux
container. Build commands lose the tag: -tags nodynamic, and
-tags renderer_integration for the integration suite.

Verified: macOS `go build ./...` and `go test ./...` pass against the stub;
full Docker image builds and starts clean on linux/arm64.
