+++ rampardos/internal/services/renderer/gopool_broadcast_linux_test.go
@@ -1,5 +1,3 @@
-//go:build mln_ffi
-
 package renderer
 
 import (
